Relocate widget and pretify code

+++ main.py
@@ -1,6 +1,7 @@
 import os
 import time
 import numpy
+import ctypes
 import shutil
 import tkinter
 import logging
@@ -15,6 +16,8 @@ from tkinter import messagebox
 from scipy.io import wavfile
 from datetime import datetime
 
+ctypes.windll.shcore.SetProcessDpiAwareness(1)
+
 logging.basicConfig(level=logging.INFO, 
                     format='%(asctime)s %(levelname)s - %(message)s',
                     datefmt='%Y-%m-%d %H:%M:%S')
@@ -23,58 +26,69 @@ logging.basicConfig(level=logging.INFO,
 class GuestbookApp:
     def __init__(self, master):
         self.master = master
+
         master.title('GUESTBOOK APP')
 
+        # init variables
         self.currently_playing = False
+        self.playing_audio_duration = 0
+        self.folder = 'master'
+        self.destination_folder = 'fd'
+        self.selected_folder = None
+
+        if not os.path.exists(self.folder):
+            os.makedirs(self.folder)
         
+        # setting windows
         screen_width = master.winfo_screenwidth()
         screen_height = master.winfo_screenheight()
 
-        x_pos = (screen_width - 400) // 2
-        y_pos = (screen_height - 300) // 2
-        master.geometry(f'400x300+{x_pos}+{y_pos}')
+        width = 550
+        height = 300
 
-        self.label = tkinter.Label(master, text='Status')
-        self.label.place(relx=0.1, rely=0.05, anchor='center')
+        x_pos = (screen_width - width) // 2
+        y_pos = (screen_height - height) // 2
+        master.geometry(f'{width}x{height}+{x_pos}+{y_pos}')
+
+        # init widgets
+        self.label = tkinter.Label(master, text='Status: Idle', justify='left')
+        self.label.place(relx=0.54, rely=0.69)
 
         self.listbox1 = tkinter.Listbox(master, height=5)
         self.listbox1.bind('<<ListboxSelect>>', self.on_select_listbox1)
-        self.listbox1.place(relx=0.1, rely=0.15, relwidth=0.4, relheight=0.7)
+        self.listbox1.place(relx=0.01, rely=0.07, relwidth=0.4, relheight=0.5)
         self.scrollbar1 = tkinter.Scrollbar(master, orient='vertical', command=self.listbox1.yview)
-        self.scrollbar1.place(relx=0.5, rely=0.15, relheight=0.7)
+        self.scrollbar1.place(relx=0.41, rely=0.07, relheight=0.5)
         self.listbox1.config(yscrollcommand=self.scrollbar1.set)
 
         self.listbox2 = tkinter.Listbox(master, height=5)
         self.listbox2.bind('<<ListboxSelect>>', self.on_select_listbox2) 
-        self.listbox2.place(relx=0.55, rely=0.15, relwidth=0.4, relheight=0.5)
+        self.listbox2.place(relx=0.55, rely=0.07, relwidth=0.4, relheight=0.5)
         self.scrollbar2 = tkinter.Scrollbar(master, orient='vertical', command=self.listbox2.yview)
-        self.scrollbar2.place(relx=0.95, rely=0.15, relheight=0.5)
+        self.scrollbar2.place(relx=0.95, rely=0.07, relheight=0.5)
         self.listbox2.config(yscrollcommand=self.scrollbar2.set)
 
-        self.folder = 'master'
         if os.path.exists(self.folder):
             files = os.listdir(self.folder)
             for file in files:
                 self.listbox1.insert(tkinter.END, file)
 
-        self.copy_button = tkinter.Button(master, text='Copy to Flashdisk', command=self.print_all_files)
-        self.copy_button.place(relx=0.4, rely=0.9, relwidth=0.3, relheight=0.08)
+        self.copy_button = tkinter.Button(master, text='Copy to Flashdisk', command=self.copy_files_in_directory)
+        self.copy_button.place(relx=0.01, rely=0.6, relwidth=0.4, relheight=0.11)
 
         self.delete_button = tkinter.Button(master, text='Delete', command=self.confirm_delete_folder)
-        self.delete_button.place(relx=0.1, rely=0.9, relwidth=0.3, relheight=0.08)
+        self.delete_button.place(relx=0.01, rely=0.72, relwidth=0.4, relheight=0.11)
 
         self.rename_button = tkinter.Button(master, text='Rename', command=self.rename_folder)
-        self.rename_button.place(relx=0.7, rely=0.9, relwidth=0.3, relheight=0.08)
-
-        self.playing_audio_duration = 0
+        self.rename_button.place(relx=0.01, rely=0.84, relwidth=0.4, relheight=0.11)
 
         self.progress_bar = ttk.Progressbar(master, orient='horizontal', length=160, mode='determinate')
-        self.progress_bar.place(relx=0.55, rely=0.7)
+        self.progress_bar.place(relx=0.55, rely=0.6, relwidth=0.4)
 
         keyboard.on_press_key('m', self.start_recording)
 
     def record_audio(self, samplerate=44100):
-        logging.info('Merekam suara... (Tekan tombol space untuk berhenti)')
+        logging.info('Recording audio (keyboard `space` to stop)')
         audio_data = []
         recording = True
 
@@ -94,7 +108,7 @@ class GuestbookApp:
                 if keyboard.is_pressed('space'):
                     recording = False
 
-        logging.info('Selesai merekam.')
+        logging.info('Finish recording')
         return numpy.concatenate(audio_data, axis=0)
 
     def save_audio(self, audio_data, folder, samplerate=44100):
@@ -109,17 +123,19 @@ class GuestbookApp:
         logging.info(f'Recording saved as {filename}')
 
     def start_recording(self, event):
-        self.label.config(text='Merekam suara...')
+        self.label.config(text='Status: Recording')
         audio_data = self.record_audio()
-        self.save_audio(audio_data, folder)
-        self.label.config(text='Rekaman selesai.')
+        self.save_audio(audio_data, self.folder)
+        self.label.config(text='Status: Idle')
 
     def on_select_listbox1(self, event):
         selected_index = self.listbox1.curselection()
+
         if selected_index:
             self.listbox2.delete(0, tkinter.END)
             self.selected_folder = self.listbox1.get(selected_index)
-            folder_path = os.path.join('master', self.selected_folder)
+            folder_path = os.path.join(self.folder, self.selected_folder)
+
             if os.path.exists(folder_path):
                 files = os.listdir(folder_path)
                 for file in files:
@@ -129,27 +145,32 @@ class GuestbookApp:
         selected_index2 = self.listbox2.curselection()
         if selected_index2 :
             selected_file = self.listbox2.get(selected_index2)
-            sound = f'{folder}/{self.selected_folder}/{selected_file}'
+            sound = f'{self.folder}/{self.selected_folder}/{selected_file}'
 
             if not self.currently_playing:
                 logging.info(f'Play sound: {sound}')
                 threading.Thread(target=self.play_audio, args=(sound,)).start()
 
-    def print_all_files(self):
-        logging.info(f"Files in 'master/{self.selected_folder}':")
-        folder_path = os.path.join(self.folder, self.selected_folder)
-        if os.path.exists(folder_path):
-            destination_folder = 'flashdisk'
-            if not os.path.exists(destination_folder):
-                os.makedirs(destination_folder)
+    def copy_files_in_directory(self):
+        if not self.selected_folder:
+            messagebox.showerror('Error', 'No folder selected.')
 
-            files = os.listdir(folder_path)
-            for file in files:
-                logging.info(f"- {file}")
+        else:
+            logging.info(f'Files in `master/{self.selected_folder}`:')
+            folder_path = os.path.join(self.folder, self.selected_folder)
+            if os.path.exists(folder_path):
+                if not os.path.exists(self.destination_folder):
+                    os.makedirs(self.destination_folder)
 
-                source = os.path.join(folder_path, file)
-                destination = os.path.join(destination_folder, file)
-                shutil.copyfile(source, destination)
+                files = os.listdir(folder_path)
+                for file in files:
+                    logging.info(f'- {file}')
+
+                    source = os.path.join(folder_path, file)
+                    destination = os.path.join(self.destination_folder, file)
+                    shutil.copyfile(source, destination)
+
+                messagebox.showinfo('Success', f'success copy files from {self.selected_folder}')
 
     def confirm_delete_folder(self):
         if self.selected_folder:
@@ -165,13 +186,17 @@ class GuestbookApp:
 
     def delete_folder(self):
         folder_path = os.path.join(self.folder, self.selected_folder)
+
         if os.path.exists(folder_path):
             shutil.rmtree(folder_path)
             self.listbox1.delete(0, tkinter.END)
             files = os.listdir(self.folder)
+
             for file in files:
                 self.listbox1.insert(tkinter.END, file)
-            logging.info(f"Folder '{self.selected_folder}' and its contents deleted.")
+
+            logging.info(f'Folder `{self.selected_folder}` and its contents deleted.')
+            messagebox.showinfo('Success', f'success delete folder {self.selected_folder}')
 
     def play_audio(self, audio_path, samplerate=44100):
         audio_file = wavfile.read(audio_path)
@@ -189,16 +214,22 @@ class GuestbookApp:
             time.sleep(0.1) 
             elapsed_time += 0.1
             progress_percentage = (elapsed_time / self.playing_audio_duration) * 100
+
+            self.label.config(text=f'Status: Playing ({elapsed_time:.1f}/{self.playing_audio_duration:.1f} seconds)')
+
             self.progress_bar['value'] = progress_percentage
             self.master.update()
 
+        self.label.config(text=f'Status: Idle')
+
     def rename_folder(self):
         selected_index = self.listbox1.curselection()
-        if selected_index:
-            selected_folder = self.listbox1.get(selected_index)
+        if not selected_index:
+            messagebox.showerror('Error', 'No folder selected.')
 
+        if selected_index:
             rename_window = tkinter.Toplevel(self.master)
-            rename_window.title("Rename Folder")
+            rename_window.title('Rename Folder')
             rename_window.overrideredirect(True)
 
             screen_width = self.master.winfo_screenwidth()
@@ -206,7 +237,7 @@ class GuestbookApp:
 
             keyboard_padding = 20
             keyboard_width = 350 + 2 * keyboard_padding
-            keyboard_height = 250 + 2 * keyboard_padding
+            keyboard_height = 350 + 2 * keyboard_padding
 
             x_pos = (screen_width - keyboard_width) // 2
             y_pos = (screen_height - keyboard_height) // 2
@@ -214,7 +245,6 @@ class GuestbookApp:
             rename_window.geometry(f"{keyboard_width}x{keyboard_height}+{x_pos}+{y_pos}")
 
             def add_char(character):
-                current_name = entry.get()
                 entry.insert(tkinter.END, character)
 
             def clear_entry():
@@ -270,10 +300,6 @@ class GuestbookApp:
             rename_window.destroy()
 
 if __name__ == '__main__':
-    folder = 'master'
-    if not os.path.exists(folder):
-        os.makedirs(folder)
-
     root = tkinter.Tk()
     app = GuestbookApp(root)
     root.mainloop()
